+++ src/Attribute/Auditable.php
@@ -11,7 +11,7 @@ namespace PrecisionSoft\Doctrine\Audit\Attribute;
 use Attribute;
 
 #[Attribute(Attribute::TARGET_CLASS)]
-final class Auditable
+class Auditable
 {
     public function __construct(
         public readonly bool $enabled = true,