+++ src/Dto/Storage/EntityDto.php
@@ -11,7 +11,7 @@ namespace PrecisionSoft\Doctrine\Audit\Dto\Storage;
 use PrecisionSoft\Doctrine\Audit\Dto\AbstractEntityDto;
 use PrecisionSoft\Doctrine\Audit\Dto\Operation;
 
-final class EntityDto extends AbstractEntityDto
+class EntityDto extends AbstractEntityDto
 {
     public function __construct(
         Operation $operation,